refactor(canvas-workspace): add structure guardrails
