+++ apps/canvas-workspace/src/renderer/src/types.ts
@@ -1,1858 +1,22 @@
-import type { PluginBridge } from '../../plugins/types';
-
-export interface CanvasNode {
-  id: string;
-  type:
-  | "file"
-  | "terminal"
-  | "frame"
-  | "group"
-  | "agent"
-  | "text"
-  | "iframe"
-  | "image"
-  | "shape"
-  | "mindmap"
-  | "reference"
-  | "dynamic-app";
-  title: string;
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-  ref?: CanvasNodeRef;
-  properties?: Record<string, WorkspaceNodePropertyValue>;
-  links?: WorkspaceNodeLink[];
-  data:
-  | FileNodeData
-  | TerminalNodeData
-  | FrameNodeData
-  | GroupNodeData
-  | AgentNodeData
-  | TextNodeData
-  | IframeNodeData
-  | ImageNodeData
-  | ShapeNodeData
-  | MindmapNodeData
-  | ReferenceNodeData
-  | DynamicAppNodeData;
-  /** Epoch millis of last mutation; used for cross-process merge. */
-  updatedAt?: number;
-}
-
-/**
- * Backing data for a `type: 'dynamic-app'` canvas node. The
- * dynamic-app plugin (gated behind the `dynamic-app` experimental
- * flag) owns the runner and the URL; we just persist enough to find
- * the right runner and render the iframe.
- */
-export interface DynamicAppNodeData {
-  /** Full loopback URL the iframe should load — `/ui/<id>` on the
-   *  plugin's shared HTTP server, with a `?v=<ts>` cache-buster
-   *  whenever the runner was restarted. */
-  url: string;
-  /** Identity hook for the persisted spec / state files and for the
-   *  plugin's IPC channels. Matches `tools.ts` and the reconciler. */
-  dynamicAppId: string;
-}
-
-export type WorkspaceNodePropertyValue =
-  | string
-  | number
-  | boolean
-  | null
-  | string[]
-  | number[]
-  | { type: 'date'; value: string }
-  | { type: 'url'; value: string }
-  | { type: 'file'; path: string }
-  | { type: 'node'; nodeId: string }
-  | { type: 'workspace-node'; workspaceId: string; nodeId: string };
-
-export interface WorkspaceNodeLink {
-  relation: string;
-  target: {
-    workspaceId?: string;
-    nodeId: string;
-  };
-  title?: string;
-  properties?: Record<string, WorkspaceNodePropertyValue>;
-}
-
-export interface WorkspaceNodeRecord {
-  schemaVersion: 1;
-  id: string;
-  type: string;
-  title?: string;
-  data: Record<string, unknown>;
-  properties?: Record<string, WorkspaceNodePropertyValue>;
-  links?: WorkspaceNodeLink[];
-  updatedAt?: number;
-  createdAt?: number;
-}
-
-export interface WorkspaceNodeListItem {
-  workspaceId?: string;
-  workspaceName?: string;
-  id: string;
-  type: string;
-  title?: string;
-  /** Friendlier label derived from the canvas node (text preview, mindmap root, ...). */
-  displayTitle?: string;
-  summary?: string;
-  tags: string[];
-  links?: WorkspaceNodeLink[];
-  updatedAt?: number;
-  createdAt?: number;
-  hasData: boolean;
-  linkCount: number;
-  /** Whether a canvas node with this id currently exists in the workspace. Undefined when not computed. */
-  onCanvas?: boolean;
-}
-
-export interface KnowledgeTagDefinition {
-  id: string;
-  name: string;
-  description?: string;
-  createdAt?: number;
-  updatedAt?: number;
-}
-
-export type CanvasNodeRef =
-  | {
-    kind: 'workspace-node';
-    workspaceId: string;
-    nodeId: string;
-  }
-  | {
-    kind: 'global-node';
-    nodeId: string;
-  };
-
-export interface FileNodeData {
-  filePath: string;
-  content: string;
-  saved?: boolean;
-  modified?: boolean;
-}
-
-export interface TerminalNodeData {
-  sessionId: string;
-  scrollback?: string;
-  cwd?: string;
-  /** Shell command to execute automatically after the terminal spawns. */
-  initialCommand?: string;
-}
-
-export interface FrameNodeData {
-  color: string;
-  label?: string;
-  agentTeamId?: string;
-  agentTeamName?: string;
-  agentTeamGoal?: string;
-  agentTeamPanelHeight?: number;
-}
-
-export interface GroupNodeData {
-  color?: string;
-  label?: string;
-  childIds?: string[];
-}
-
-export interface AgentNodeData {
-  sessionId: string;
-  scrollback?: string;
-  cwd?: string;
-  agentType: string;
-  status?: 'idle' | 'running' | 'done' | 'error';
-  agentArgs?: string;
-  /**
-   * When true, launch the agent in unrestricted mode:
-   *   - claude-code → adds `--dangerously-skip-permissions`
-   *   - codex       → adds `--dangerously-bypass-approvals-and-sandbox`
-   * Other agent types ignore this flag.
-   */
-  dangerousMode?: boolean;
-  /** Short prompt passed directly as a CLI argument. */
-  inlinePrompt?: string;
-  /** Relative path to a prompt file in cwd for long prompts. */
-  promptFile?: string;
-  /**
-   * Snapshot of the prompt the user supplied at launch time. `inlinePrompt`
-   * is cleared after being sent to the CLI to prevent re-sending on the
-   * next spawn, but the Restart view needs the original text to show what
-   * the previous session was started with.
-   */
-  lastInitPrompt?: string;
-  /**
-   * Current AgentNodeBody view ('setup' | 'running' | 'restart'). Persisted
-   * so the outer `CanvasNodeView` header can render the status pill that
-   * matches the body — and so a cold reload mid-running session can be
-   * distinguished from an active live PTY (the body rewrites this on
-   * mount based on the cold-reload heuristic). Optional because legacy
-   * nodes won't have it; consumers should fall back to a derived value.
-   */
-  viewMode?: 'setup' | 'running' | 'restart';
-  /**
-   * Caller-supplied session id for Claude Code. Claude accepts this on first
-   * spawn via `--session-id <uuid>` and later via `--resume <uuid>`.
-   */
-  cliSessionId?: string;
-  /**
-   * Codex session id captured after first launch from Codex's local session
-   * index. Codex does not accept caller-supplied ids on first spawn, but it
-   * does support `codex resume <id>` once the created id is known.
-   */
-  codexSessionId?: string;
-  /**
-   * Short host marker appended to the initial Codex prompt so the renderer can
-   * bind the created Codex thread id from local metadata after launch.
-   */
-  codexSessionMarker?: string;
-  agentTeamAutoResume?: {
-    sessionKey?: string;
-    attempts?: number;
-    lastAttemptAt?: number;
-  };
-  /** Team-managed node should launch an idle CLI session before task dispatch. */
-  agentTeamWarmup?: boolean;
-  /** The warmup CLI has produced output and is ready to receive team input. */
-  agentTeamWarmupReady?: boolean;
-  agentTeamId?: string;
-  agentTeamAgentId?: string;
-  agentTeamRole?: 'lead' | 'teammate';
-}
-
-/**
- * TLDRAW-style free-form text label on the canvas.
- *
- * Content is markdown (supports **bold**, *italic*, ++underline++, lists,
- * headings, etc.). Colors are applied via inline styles so they persist
- * across reloads. `backgroundColor: 'transparent'` renders a chrome-free
- * label.
- *
- * `autoSize` controls how the wrapper dimensions are interpreted:
- *  - `true` (default): width/height track the rendered content via CSS
- *    `max-content`. Useful while typing so the node hugs its text.
- *  - `false`: the user has manually resized the node; `width`/`height`
- *    become a hard frame and long text wraps within. Set by the resize
- *    drag handles.
- */
-export interface TextNodeData {
-  content: string;
-  textColor: string;
-  backgroundColor: string;
-  /** Optional font size in px; defaults to 18 when unset. */
-  fontSize?: number;
-  /** When false, the user has dragged a resize handle — respect width/height. */
-  autoSize?: boolean;
-}
-
-/**
- * Embeds an external web page or renders raw HTML.
- *
- * `mode: 'url'` (default) loads a remote page via Electron `<webview>`.
- * `mode: 'html'` renders user-supplied HTML in a sandboxed `<iframe srcdoc>`.
- *
- * When `artifactId` is set, the rendered HTML is sourced from the
- * workspace's artifact store (current version) instead of `html`. This
- * lets the same artifact back both an in-chat preview and a pinned
- * canvas node — iterating the artifact updates both.
- *
- * Some sites block embedding via X-Frame-Options / CSP frame-ancestors —
- * those will fail to render and the user can click "Open externally" to
- * escape into the system browser.
- */
-export interface IframeNodeData {
-  /** Full URL (including protocol) to load in the iframe. Empty = show URL input. `about:blank` opens a blank page. */
-  url: string;
-  /** Raw HTML content to render when `mode` is `'html'` or `'ai'`. */
-  html?: string;
-  /** `'url'` embeds a remote page; `'html'` renders raw HTML; `'ai'` generates HTML from a prompt; `'artifact'` pulls from the artifact store. */
-  mode?: 'url' | 'html' | 'ai' | 'artifact';
-  /** The prompt used to generate HTML when `mode` is `'ai'`. */
-  prompt?: string;
-  /** Last page title reported by the embedded webview for URL mode. */
-  pageTitle?: string;
-  /** Last favicon URL reported by the embedded webview for URL mode. */
-  faviconUrl?: string;
-  /** When set, content is sourced from `artifacts.get(artifactId)`. */
-  artifactId?: string;
-}
-
-/**
- * A free-form image pinned to the canvas. Created by pasting (Ctrl/Cmd+V)
- * image data onto the canvas; the bytes are persisted to the workspace
- * images directory via `file:saveImage` and referenced by absolute path.
- *
- * Rendered with no header/border chrome so the picture is the whole node.
- * The entire body acts as the drag handle (mirroring TextNodeBody).
- */
-export interface ImageNodeData {
-  /** Absolute path on disk to the saved image file. */
-  filePath: string;
-}
-
-/**
- * A primitive geometric shape drawn on the canvas. Rendered as an inline
- * SVG that fills the node box, so the same storage powers both rect and
- * ellipse (and any future primitives). Colors are stored as hex strings
- * or "transparent"; stroke width is in CSS pixels at scale=1.
- *
- * `text` is an optional free-form label centered inside the shape. It
- * is rendered via an HTML `contenteditable` overlay (not SVG text) so it
- * wraps like any other DOM text and inherits the canvas's font stack.
- */
-export interface ShapeNodeData {
-  kind: "rect" | "rounded-rect" | "ellipse" | "triangle" | "diamond" | "hexagon" | "star";
-  fill: string;
-  stroke: string;
-  strokeWidth: number;
-  text?: string;
-  textColor?: string;
-  /** Font size in px. Defaults to 16 when unset. */
-  fontSize?: number;
-}
-
-/**
- * Heptabase-style mindmap card.
- *
- * The whole mindmap lives inside ONE canvas node — branches and children
- * are stored in a recursive `MindmapTopic` tree under `data.root`, not as
- * separate `CanvasNode`s connected by `CanvasEdge`s. This keeps the
- * structural relationships (parent/child) distinct from free-form canvas
- * connections drawn by the user between different nodes.
- *
- * Layout is deterministic: given the tree, `layoutMindmap` produces the
- * on-screen positions, so we never persist x/y for individual topics —
- * only the tree, which is cheap to round-trip through JSON storage.
- */
-export interface MindmapTopic {
-  /** Stable id — unique within the containing mindmap only. */
-  id: string;
-  text: string;
-  children: MindmapTopic[];
-  /** When true, the subtree is hidden (and its descendants are skipped
-   *  by layout). Applies only to non-root topics. */
-  collapsed?: boolean;
-  /** Optional per-topic color override (hex). Defaults to a branch color
-   *  derived from the topic's root-child index. */
-  color?: string;
-}
-
-export interface MindmapNodeData {
-  root: MindmapTopic;
-  /** Layout direction. v1 supports `'right'` (root on the left, children
-   *  growing rightward); other values are reserved for future modes. */
-  layout: 'right';
-  /** Bumped on every topic add/remove/rename — lets external observers
-   *  (canvas-cli, agent) diff mindmaps without structural equality. */
-  rev?: number;
-}
-
-export interface ReferenceNodeData {
-  titleSnapshot?: string;
-  typeSnapshot?: Exclude<CanvasNode['type'], 'reference'>;
-  workspaceNameSnapshot?: string;
-}
-
-export interface CanvasTransform {
-  x: number;
-  y: number;
-  scale: number;
-}
-
-/**
- * One end of a `CanvasEdge`.
- *
- * `node`  → the endpoint tracks a node's anchor; the actual screen position
- *           is computed each render from the node's x/y/width/height.
- * `point` → the endpoint is a free canvas coordinate (no node binding).
- *           Used when the user drags an endpoint into empty space, or when
- *           a bound node is deleted and we degrade instead of cascade-delete.
- */
-export type EdgeAnchor = 'top' | 'right' | 'bottom' | 'left' | 'auto';
-
-export type EdgeEndpoint =
-  | { kind: 'node'; nodeId: string; anchor?: EdgeAnchor }
-  | { kind: 'point'; x: number; y: number };
-
-export type EdgeArrowCap = 'none' | 'triangle' | 'arrow' | 'dot' | 'bar';
-
-export interface EdgeStroke {
-  color?: string;
-  width?: number;
-  style?: 'solid' | 'dashed' | 'dotted';
-}
-
-/**
- * A connection drawn between two endpoints on the canvas.
- *
- * Edges are tldraw-style: they are first-class shapes, not pure
- * "relations between nodes". Either endpoint can be free
- * (`{ kind: 'point', x, y }`) so users can draw arrows into empty
- * space; when a bound node is deleted, the renderer degrades the
- * affected endpoint to a free point rather than removing the edge.
- *
- * `bend` is a single signed scalar: 0 = straight, +N = pixels of
- * perpendicular offset from the source→target midpoint (toward the
- * normal's positive direction), -N = the other side. This keeps the
- * geometry representable as a single-control quadratic Bezier and
- * matches tldraw's arrow storage.
- *
- * `payload` is a free-form dictionary keyed by the edge's `kind` —
- * e.g. `kind: 'context'` edges can stash prompt-injection hints for
- * the canvas agent, `kind: 'data'` edges a transformer config, etc.
- * Everything stored must be JSON-safe; it round-trips through
- * canvas-store as-is.
- */
-export interface CanvasEdge {
-  id: string;
-  source: EdgeEndpoint;
-  target: EdgeEndpoint;
-  bend?: number;
-  arrowHead?: EdgeArrowCap;
-  arrowTail?: EdgeArrowCap;
-  stroke?: EdgeStroke;
-  label?: string;
-  kind?: string;
-  payload?: Record<string, unknown>;
-  /** Epoch millis of last mutation; used for cross-process merge. */
-  updatedAt?: number;
-}
-
-export interface CanvasSaveData {
-  nodes: CanvasNode[];
-  /** Connections between nodes. Optional for backwards compatibility with
-   *  older canvas.json files that pre-date connections. */
-  edges?: CanvasEdge[];
-  transform: CanvasTransform;
-  savedAt: string;
-}
-
-export interface DirEntry {
-  name: string;
-  type: 'file' | 'dir';
-  children?: DirEntry[];
-}
-
-export interface FileApi {
-  createNote: (
-    workspaceId?: string,
-    name?: string
-  ) => Promise<{ ok: boolean; filePath?: string; fileName?: string; error?: string }>;
-  read: (
-    filePath: string
-  ) => Promise<{ ok: boolean; content?: string; error?: string }>;
-  write: (
-    filePath: string,
-    content: string
-  ) => Promise<{ ok: boolean; error?: string }>;
-  listDir: (
-    dirPath: string,
-    maxDepth?: number
-  ) => Promise<{ ok: boolean; entries?: DirEntry[]; error?: string }>;
-  openDialog: () => Promise<{
-    ok: boolean;
-    canceled?: boolean;
-    filePath?: string;
-    fileName?: string;
-    content?: string;
-    error?: string;
-  }>;
-  saveAsDialog: (
-    defaultName: string,
-    content: string
-  ) => Promise<{
-    ok: boolean;
-    canceled?: boolean;
-    filePath?: string;
-    fileName?: string;
-    error?: string;
-  }>;
-  saveImage: (
-    workspaceId: string | undefined,
-    data: string,
-    ext?: string
-  ) => Promise<{ ok: boolean; filePath?: string; fileName?: string; error?: string }>;
-  exportImage: (
-    defaultName: string,
-    data: string,
-    ext?: string
-  ) => Promise<{ ok: boolean; canceled?: boolean; filePath?: string; fileName?: string; error?: string }>;
-  onChanged: (callback: (filePath: string, content: string) => void) => () => void;
-}
-
-export interface DialogApi {
-  openFolder: () => Promise<{ ok: boolean; canceled?: boolean; folderPath?: string; error?: string }>;
-}
-
-export interface SkillTargetResult {
-  path: string;
-  ok: boolean;
-  error?: string;
-}
-
-export interface SkillsInstallResult {
-  ok: boolean;
-  skillsInstalled: boolean;
-  results: SkillTargetResult[];
-  cliInstalled: boolean;
-  manualCommand?: string | null;
-  cliError?: string | null;
-  error?: string;
-}
-
-export interface SkillsStatusResult {
-  installed: boolean;
-  results: SkillTargetResult[];
-  legacyDirs: string[];
-}
-
-export interface SkillsCleanupResult {
-  ok: boolean;
-  results: SkillTargetResult[];
-}
-
-export interface SkillsApi {
-  install: () => Promise<SkillsInstallResult>;
-  status: () => Promise<SkillsStatusResult>;
-  cleanupLegacy: () => Promise<SkillsCleanupResult>;
-}
-
-export interface ChatImageAttachment {
-  id: string;
-  path: string;
-  fileName?: string;
-  mimeType?: string;
-}
-
-export type AgentScope =
-  | { kind: 'workspace'; workspaceId: string }
-  | { kind: 'global' };
-
-export interface AgentScopeRef {
-  scope: AgentScope;
-}
-
-export interface AgentChatToolCall {
-  id: number;
-  name: string;
-  args?: unknown;
-  status: 'running' | 'done';
-  result?: string;
-  toolCallId?: string;
-  partialInput?: string;
-  inputStreaming?: boolean;
-  streamedContent?: string;
-  streamedDone?: boolean;
-}
-
-export interface AgentDebugTraceNodeRef {
-  id: string;
-  title: string;
-  type: string;
-  workspaceId?: string;
-  workspaceName?: string;
-  contentChars?: number;
-  source?: 'selected' | 'read_node' | 'read_context';
-}
-
-export interface AgentDebugTraceContextRead {
-  workspaceId?: string;
-  workspaceName?: string;
-  detail?: string;
-  nodeCount?: number;
-  resultChars?: number;
-}
-
-export interface AgentDebugTraceToolCall {
-  name: string;
-  toolCallId?: string;
-  status: 'running' | 'done' | 'error';
-  argsPreview?: string;
-  resultSummary?: string;
-  startedAt?: number;
-  finishedAt?: number;
-  durationMs?: number;
-  readNodes?: AgentDebugTraceNodeRef[];
-  contextRead?: AgentDebugTraceContextRead;
-}
-
-export interface AgentDebugTraceMessageSnapshot {
-  systemPrompt: string;
-  systemPromptChars: number;
-  messagesJson: string;
-  messagesChars: number;
-  messageCount: number;
-  limitChars: number;
-  truncated?: boolean;
-}
-
-export interface AgentDebugTrace {
-  sessionId: string;
-  runId: string;
-  turnId: string;
-  createdAt: number;
-  startedAt: number;
-  finishedAt?: number;
-  durationMs?: number;
-  debugUrl?: string;
-  request: {
-    userPromptPreview: string;
-    attachmentCount: number;
-    executionMode?: 'auto' | 'ask';
-    scope?: 'current_canvas' | 'selected_nodes';
-    quickAction?: string;
-    selectedNodes: AgentDebugTraceNodeRef[];
-    mentionedCanvases: Array<{ id: string; name: string }>;
-    workspace?: {
-      id: string;
-      name: string;
-      nodeCount: number;
-    };
-  };
-  prompt: {
-    systemPromptPreview: string;
-    systemPromptChars: number;
-    currentCanvasSummaryPreview?: string;
-    currentCanvasSummaryChars?: number;
-  };
-  messageSnapshot?: AgentDebugTraceMessageSnapshot;
-  model?: {
-    provider?: string;
-    model?: string;
-    modelType?: string;
-  };
-  toolCalls: AgentDebugTraceToolCall[];
-  readNodes: AgentDebugTraceNodeRef[];
-  contextReads: AgentDebugTraceContextRead[];
-  truncated?: boolean;
-}
-
-export interface AgentDebugRunSummary {
-  workspaceId: string;
-  workspaceName: string;
-  sessionId: string;
-  runId: string;
-  turnId: string;
-  messageIndex: number;
-  startedAt: number;
-  durationMs?: number;
-  userPromptPreview: string;
-  assistantPreview: string;
-  toolCount: number;
-  readNodeCount: number;
-  modelLabel?: string;
-  isCurrent: boolean;
-}
-
-export interface AgentDebugRunDetail extends AgentDebugRunSummary {
-  userMessage?: AgentChatMessage;
-  assistantMessage?: AgentChatMessage;
-  trace: AgentDebugTrace;
-}
-
-export interface AgentChatMessage {
-  role: 'user' | 'assistant';
-  content: string;
-  timestamp: number;
-  attachments?: ChatImageAttachment[];
-  toolCalls?: AgentChatToolCall[];
-  // Stable identifier of the agent turn that produced this message.
-  // Plugins (e.g. devtools) look up turn-scoped data — such as the
-  // captured debug trace — by this id via their own storage.
-  runId?: string;
-}
-
-export interface AgentContextNodeRef {
-  id: string;
-  title: string;
-  type: CanvasNode['type'];
-  /**
-   * Owning workspace of the node. Required for global-scope chat (Nodes /
-   * Graph knowledge assistant), where there is no bound canvas and the agent
-   * must pass an explicit workspaceId to read the node. Omitted for the
-   * workspace-scoped canvas ChatPanel, where the scope already pins it.
-   */
-  workspaceId?: string;
-}
-
-export interface AgentContextTagRef {
-  name: string;
-  /**
-   * Workspaces where this tag occurs. In global-scope chat the agent needs an
-   * explicit workspaceId to query a tag's members (`workspace_node_list`).
-   */
-  workspaceIds?: string[];
-}
-
-export interface AgentContextCanvasRef {
-  id: string;
-  name: string;
-}
-
-export interface AgentRequestContext {
-  executionMode?: 'auto' | 'ask';
-  scope?: 'current_canvas' | 'selected_nodes';
-  selectedNodes?: AgentContextNodeRef[];
-  /** Tags the user scoped the turn to (global Nodes/Graph assistant). */
-  tags?: AgentContextTagRef[];
-  /** Whole canvases the user scoped the turn to (global assistant). */
-  canvases?: AgentContextCanvasRef[];
-  quickAction?: string;
-}
-
-export interface AgentSessionInfo {
-  sessionId: string;
-  date: string;
-  messageCount: number;
-  isCurrent: boolean;
-  preview?: string;
-}
-
-export interface CrossWorkspaceSessionGroup {
-  workspaceId: string;
-  workspaceName: string;
-  sessions: AgentSessionInfo[];
-}
-
-/** One hit from the session title search behind the @-mention popup. */
-export interface SessionSearchHit {
-  sessionId: string;
-  workspaceId: string;
-  workspaceName: string;
-  date: string;
-  isCurrent: boolean;
-  messageCount: number;
-  preview: string;
-}
-
-
-export type CanvasModelProviderType = 'openai' | 'claude';
-
-export interface CanvasModelOption {
-  name: string;
-  provider_type?: CanvasModelProviderType;
-  model?: string;
-  base_url?: string;
-  api_key_env?: string;
-  headers?: Record<string, string>;
-}
-
-export interface CanvasProviderModel {
-  id: string;
-  name?: string;
-}
-
-export interface CanvasModelProviderConfig {
-  id: string;
-  name: string;
-  provider_type?: CanvasModelProviderType;
-  base_url?: string;
-  api_key_env?: string;
-  api_key?: string;
-  headers?: Record<string, string>;
-  models?: CanvasProviderModel[];
-}
-
-export interface CanvasModelConfig {
-  current_provider?: string;
-  current_model?: string;
-  provider_type?: CanvasModelProviderType;
-  model?: string;
-  base_url?: string;
-  api_key_env?: string;
-  headers?: Record<string, string>;
-  options?: CanvasModelOption[];
-  providers?: CanvasModelProviderConfig[];
-}
-
-export interface CanvasModelProviderStatus {
-  id: string;
-  name: string;
-  provider_type: CanvasModelProviderType;
-  base_url?: string;
-  api_key_env?: string;
-  apiKeyPresent: boolean;
-  apiKeyLength?: number;
-  headers?: Record<string, string>;
-  models: CanvasProviderModel[];
-}
-
-export interface CanvasModelStatus {
-  path: string;
-  currentProvider?: string;
-  currentModel?: string;
-  providerType: CanvasModelProviderType;
-  resolvedModel: string;
-  resolvedBaseURL?: string;
-  resolvedApiKeyEnv?: string;
-  apiKeyPresent: boolean;
-  options: CanvasModelOption[];
-  providers: CanvasModelProviderStatus[];
-}
-
-export type PromptPreset = 'concise' | 'balanced' | 'detailed';
-
-export interface PromptProfile {
-  preset: PromptPreset;
-  customPrompt: string;
-}
-
-export interface PromptProfileStatus extends PromptProfile {
-  path: string;
-}
-
-export interface PromptProfileApi {
-  get: () => Promise<{ ok: boolean; profile?: PromptProfileStatus; error?: string }>;
-  save: (
-    profile: Partial<PromptProfile>,
-  ) => Promise<{ ok: boolean; profile?: PromptProfileStatus; error?: string }>;
-  reset: () => Promise<{ ok: boolean; profile?: PromptProfileStatus; error?: string }>;
-}
-
-export interface CanvasModelApi {
-  status: () => Promise<{ ok: boolean; status?: CanvasModelStatus; error?: string }>;
-  saveConfig: (config: CanvasModelConfig) => Promise<{ ok: boolean; status?: CanvasModelStatus; error?: string }>;
-  upsertProvider: (provider: CanvasModelProviderConfig) => Promise<{ ok: boolean; status?: CanvasModelStatus; error?: string }>;
-  removeProvider: (providerId: string) => Promise<{ ok: boolean; status?: CanvasModelStatus; error?: string }>;
-  fetchModels: (
-    providerId?: string,
-    provider?: CanvasModelProviderConfig,
-  ) => Promise<{ ok: boolean; models?: CanvasProviderModel[]; error?: string }>;
-  upsertOption: (
-    option: CanvasModelOption,
-    setCurrent?: boolean,
-  ) => Promise<{ ok: boolean; status?: CanvasModelStatus; error?: string }>;
-  setCurrent: (name?: string, providerId?: string) => Promise<{ ok: boolean; status?: CanvasModelStatus; error?: string }>;
-  removeOption: (name: string) => Promise<{ ok: boolean; status?: CanvasModelStatus; error?: string }>;
-  reset: () => Promise<{ ok: boolean; status?: CanvasModelStatus; error?: string }>;
-}
-
-export type BuiltInToolCredentialId = 'openai' | 'gemini' | 'tavily';
-
-export interface BuiltInToolCredentialStatus {
-  id: BuiltInToolCredentialId;
-  name: string;
-  description: string;
-  envKey: string;
-  baseUrlEnvKey: string;
-  defaultBaseUrl: string;
-  tools: string[];
-  apiKeyPresent: boolean;
-  apiKeyLength?: number;
-  source: 'stored' | 'env' | 'missing';
-  baseUrl: string;
-  baseUrlSource: 'stored' | 'env' | 'default';
-}
-
-export interface BuiltInToolsConfigStatus {
-  path: string;
-  credentials: BuiltInToolCredentialStatus[];
-}
-
-export interface BuiltInToolsConfigApi {
-  status: () => Promise<{ ok: boolean; status?: BuiltInToolsConfigStatus; error?: string }>;
-  setCredential: (
-    id: BuiltInToolCredentialId,
-    input: { apiKey?: string; baseUrl?: string },
-  ) => Promise<{ ok: boolean; status?: BuiltInToolsConfigStatus; error?: string }>;
-  clearCredential: (
-    id: BuiltInToolCredentialId,
-  ) => Promise<{ ok: boolean; status?: BuiltInToolsConfigStatus; error?: string }>;
-}
-
-// --- User-configurable Skills & MCP (global + per-workspace scope) ---
-
-export type CanvasConfigScope =
-  | { level: 'global' }
-  | { level: 'workspace'; workspaceId: string };
-
-export type CanvasSkillSourceName =
-  | 'canvas'
-  | 'pulse-coder'
-  | 'agents'
-  | 'coder'
-  | 'claude'
-  | 'codex';
-
-export interface CanvasSkillEntry {
-  name: string;
-  description: string;
-  body: string;
-  scope: 'global' | 'workspace';
-  path: string;
-  source: CanvasSkillSourceName;
-  /** False for skills owned by other agent tools — Edit/Delete are hidden. */
-  writable: boolean;
-}
-
-export interface CanvasSkillsStatus {
-  scope: 'global' | 'workspace';
-  dir: string;
-  skills: CanvasSkillEntry[];
-}
-
-export interface CanvasSkillInput {
-  name: string;
-  description: string;
-  body: string;
-  originalName?: string;
-}
-
-export interface CanvasSkillImportEntry {
-  name: string;
-  status: 'imported' | 'replaced' | 'skipped';
-  reason?: string;
-}
-
-export interface CanvasSkillsApi {
-  list: (scope: CanvasConfigScope) => Promise<{ ok: boolean; status?: CanvasSkillsStatus; error?: string }>;
-  upsert: (
-    scope: CanvasConfigScope,
-    skill: CanvasSkillInput,
-  ) => Promise<{ ok: boolean; status?: CanvasSkillsStatus; error?: string }>;
-  remove: (
-    scope: CanvasConfigScope,
-    name: string,
-  ) => Promise<{ ok: boolean; status?: CanvasSkillsStatus; error?: string }>;
-  importZip: (
-    scope: CanvasConfigScope,
-    bytes: ArrayBuffer,
-  ) => Promise<{
-    ok: boolean;
-    status?: CanvasSkillsStatus;
-    entries?: CanvasSkillImportEntry[];
-    error?: string;
-  }>;
-  importMd: (
-    scope: CanvasConfigScope,
-    text: string,
-  ) => Promise<{
-    ok: boolean;
-    status?: CanvasSkillsStatus;
-    name?: string;
-    result?: 'imported' | 'replaced';
-    error?: string;
-  }>;
-  importUrl: (
-    scope: CanvasConfigScope,
-    url: string,
-  ) => Promise<{
-    ok: boolean;
-    status?: CanvasSkillsStatus;
-    /** Tells the caller which underlying importer ran, so it can pick the right toast. */
-    kind?: 'md' | 'zip';
-    name?: string;
-    result?: 'imported' | 'replaced';
-    entries?: CanvasSkillImportEntry[];
-    error?: string;
-  }>;
-}
-
-export type CanvasMcpTransport = 'http' | 'sse' | 'stdio';
-
-export interface CanvasMcpServer {
-  name: string;
-  transport: CanvasMcpTransport;
-  url?: string;
-  headers?: Record<string, string>;
-  command?: string;
-  args?: string[];
-  env?: Record<string, string>;
-  cwd?: string;
-  deferTools?: boolean;
-  /** Bare tool names the user has turned off; the engine skips registering these. */
-  disabledTools?: string[];
-}
-
-/** One tool exposed by a connected MCP server, with its enabled state. */
-export interface CanvasMcpToolInfo {
-  name: string;
-  description?: string;
-  enabled: boolean;
-}
-
-export type CanvasMcpServerHealth =
-  | { ok: true; toolCount: number; tools?: CanvasMcpToolInfo[] }
-  | { ok: false; error: string };
-
-export interface CanvasMcpStatus {
-  scope: 'global' | 'workspace';
-  path: string;
-  servers: CanvasMcpServer[];
-  /**
-   * Per-server connection health from the engine's MCP plugin. Servers
-   * absent from this map have never been loaded by an active agent yet.
-   */
-  statuses?: Record<string, CanvasMcpServerHealth>;
-}
-
-export interface CanvasMcpImportEntry {
-  name: string;
-  status: 'added' | 'replaced' | 'skipped';
-  reason?: string;
-}
-
-export interface CanvasMcpApi {
-  list: (scope: CanvasConfigScope) => Promise<{ ok: boolean; status?: CanvasMcpStatus; error?: string }>;
-  upsert: (
-    scope: CanvasConfigScope,
-    server: CanvasMcpServer,
-    originalName?: string,
-  ) => Promise<{ ok: boolean; status?: CanvasMcpStatus; error?: string }>;
-  remove: (
-    scope: CanvasConfigScope,
-    name: string,
-  ) => Promise<{ ok: boolean; status?: CanvasMcpStatus; error?: string }>;
-  importJson: (
-    scope: CanvasConfigScope,
-    json: string,
-  ) => Promise<{
-    ok: boolean;
-    status?: CanvasMcpStatus;
-    entries?: CanvasMcpImportEntry[];
-    error?: string;
-  }>;
-  setToolEnabled: (
-    scope: CanvasConfigScope,
-    name: string,
-    tool: string,
-    enabled: boolean,
-  ) => Promise<{ ok: boolean; status?: CanvasMcpStatus; error?: string }>;
-}
-
-export interface AgentApi {
-  chat: (
-    scopeRef: AgentScopeRef,
-    message: string,
-    mentionedWorkspaceIds?: string[],
-    requestContext?: AgentRequestContext,
-    attachments?: ChatImageAttachment[]
-  ) => Promise<{ ok: boolean; sessionId?: string; error?: string }>;
-  onTextDelta: (
-    sessionId: string,
-    callback: (delta: string) => void
-  ) => () => void;
-  onChatComplete: (
-    sessionId: string,
-    callback: (result: { ok: boolean; response?: string; runId?: string; error?: string }) => void
-  ) => () => void;
-  onToolCall: (
-    sessionId: string,
-    callback: (data: { name: string; args: any; toolCallId?: string }) => void
-  ) => () => void;
-  onToolResult: (
-    sessionId: string,
-    callback: (data: { name: string; result: string; toolCallId?: string }) => void
-  ) => () => void;
-  /** Tool-input streaming — fired when LLM starts emitting tool arguments. */
-  onToolInputStart: (
-    sessionId: string,
-    callback: (data: { id: string; toolName: string }) => void
-  ) => () => void;
-  /** Each chunk of raw tool argument JSON. `id` matches `toolCallId` on the final tool-call. */
-  onToolInputDelta: (
-    sessionId: string,
-    callback: (data: { id: string; delta: string }) => void
-  ) => () => void;
-  onToolInputEnd: (
-    sessionId: string,
-    callback: (data: { id: string }) => void
-  ) => () => void;
-  /**
-   * Subscribe to side-channel visual stream chunks emitted by the
-   * `visual_render` tool when the upstream LLM/provider doesn't stream
-   * tool-call arguments. The tool itself chunks its final content and
-   * broadcasts updates keyed by `toolCallId`, which the renderer matches
-   * to the corresponding ToolCallStatus.
-   */
-  onVisualStream: (
-    callback: (data: {
-      workspaceId: string;
-      toolCallId: string;
-      content: string;
-      done?: boolean;
-    }) => void
-  ) => () => void;
-  onClarifyRequest: (
-    sessionId: string,
-    callback: (data: { id: string; question: string; context?: string }) => void
-  ) => () => void;
-  answerClarification: (
-    sessionId: string,
-    requestId: string,
-    answer: string
-  ) => Promise<{ ok: boolean; error?: string }>;
-  abort: (sessionId: string) => Promise<{ ok: boolean; error?: string }>;
-  getStatus: (
-    scopeRef: AgentScopeRef
-  ) => Promise<{ ok: boolean; active: boolean; messageCount: number }>;
-  listSkills: (
-    scopeRef: AgentScopeRef
-  ) => Promise<{ ok: boolean; skills?: Array<{ name: string; description: string }>; error?: string }>;
-  getHistory: (
-    scopeRef: AgentScopeRef
-  ) => Promise<{ ok: boolean; messages?: AgentChatMessage[] }>;
-  listSessions: (
-    scopeRef: AgentScopeRef
-  ) => Promise<{ ok: boolean; sessions?: AgentSessionInfo[]; error?: string }>;
-  newSession: (
-    scopeRef: AgentScopeRef
-  ) => Promise<{ ok: boolean; error?: string }>;
-  rewindMessages: (
-    scopeRef: AgentScopeRef,
-    fromIndex: number,
-  ) => Promise<{ ok: boolean; error?: string }>;
-  loadSession: (
-    scopeRef: AgentScopeRef,
-    sessionId: string
-  ) => Promise<{ ok: boolean; messages?: AgentChatMessage[]; error?: string }>;
-  listAllSessions: (
-    workspaceNames: Record<string, string>,
-  ) => Promise<{ ok: boolean; groups?: CrossWorkspaceSessionGroup[]; error?: string }>;
-  /** Keyword search over stored session message content (for the @-mention popup). */
-  searchSessions: (
-    query: string,
-    limit?: number,
-  ) => Promise<{ ok: boolean; hits?: SessionSearchHit[]; error?: string }>;
-  /** Current session id for a scope (live agent, falling back to disk). */
-  getCurrentSession: (
-    scopeRef: AgentScopeRef
-  ) => Promise<{ ok: boolean; sessionId?: string | null; error?: string }>;
-  loadCrossWorkspaceSession: (
-    targetWorkspaceId: string,
-    sourceWorkspaceId: string,
-    sessionId: string,
-  ) => Promise<{ ok: boolean; messages?: AgentChatMessage[]; error?: string }>;
-  activate: (workspaceId: string) => Promise<{ ok: boolean; error?: string }>;
-  deactivate: (workspaceId: string) => Promise<{ ok: boolean; error?: string }>;
-  addImageToCanvas: (
-    workspaceId: string,
-    imagePath: string,
-    title?: string
-  ) => Promise<{ ok: boolean; nodeId?: string; error?: string }>;
-  streamWorkspaceDoc: (payload: {
-    workspaceName: string;
-    intent: string;
-    currentContent?: string;
-  }) => Promise<{ ok: boolean; requestId?: string; error?: string }>;
-  onWorkspaceDocDelta: (requestId: string, callback: (delta: string) => void) => () => void;
-  onWorkspaceDocComplete: (
-    requestId: string,
-    callback: (result: { ok: boolean; content?: string; error?: string }) => void,
-  ) => () => void;
-}
-
-export type AgentTeamStatus =
-  | 'planning'
-  | 'waiting_approval'
-  | 'running'
-  | 'reviewing'
-  | 'paused'
-  | 'round_checkpoint'
-  | 'completed'
-  | 'failed';
-
-export type AgentTeamAgentStatus =
-  | 'idle'
-  | 'running'
-  | 'needs_input'
-  | 'blocked'
-  | 'done'
-  | 'error'
-  | 'stopped';
-
-export type AgentTeamTaskStatus =
-  | 'todo'
-  | 'in_progress'
-  | 'needs_input'
-  | 'needs_review'
-  | 'blocked'
-  | 'done'
-  | 'failed';
-
-export interface AgentTeamRecord {
-  id: string;
-  name: string;
-  goal: string;
-  status: AgentTeamStatus;
-  leadAgentId?: string;
-  createdAt: number;
-  updatedAt: number;
-  metadata?: Record<string, unknown>;
-}
-
-export interface AgentTeamAgentRecord {
-  id: string;
-  teamId: string;
-  role: 'lead' | 'teammate';
-  name: string;
-  status: AgentTeamAgentStatus;
-  cwd?: string;
-  currentTaskId?: string;
-  sessionRef?: {
-    sessionId: string;
-    provider: string;
-    displayName?: string;
-    metadata?: Record<string, unknown>;
-  };
-  createdAt: number;
-  updatedAt: number;
-  metadata?: Record<string, unknown>;
-}
-
-export interface AgentTeamTaskRecord {
-  id: string;
-  teamId: string;
-  title: string;
-  description: string;
-  status: AgentTeamTaskStatus;
-  ownerAgentId?: string;
-  deps: string[];
-  result?: string;
-  blockedReason?: string;
-  createdBy: string;
-  createdAt: number;
-  updatedAt: number;
-  metadata?: Record<string, unknown>;
-}
-
-export interface AgentTeamHumanGateRecord {
-  id: string;
-  teamId: string;
-  taskId?: string;
-  agentId?: string;
-  reason: string;
-  prompt: string;
-  status: 'open' | 'answered' | 'cancelled';
-  answer?: string;
-  createdAt: number;
-  updatedAt: number;
-  metadata?: Record<string, unknown>;
-}
-
-export interface AgentTeamArtifactRecord {
-  id: string;
-  teamId: string;
-  taskId?: string;
-  agentId?: string;
-  kind: string;
-  title: string;
-  uri?: string;
-  summary?: string;
-  createdAt: number;
-  metadata?: Record<string, unknown>;
-}
-
-export interface AgentTeamEventRecord {
-  id: string;
-  teamId: string;
-  type: string;
-  timestamp: number;
-  actor: string;
-  payload: Record<string, unknown>;
-}
-
-export interface AgentTeamMessageRecord {
-  id: string;
-  teamId: string;
-  from: string;
-  to: string;
-  type: string;
-  content: string;
-  taskId?: string;
-  createdAt: number;
-  readAt?: number;
-  metadata?: Record<string, unknown>;
-}
-
-export interface AgentTeamRuntimeSnapshot {
-  team: AgentTeamRecord;
-  agents: AgentTeamAgentRecord[];
-  tasks: AgentTeamTaskRecord[];
-  artifacts: AgentTeamArtifactRecord[];
-  humanGates: AgentTeamHumanGateRecord[];
-  events: AgentTeamEventRecord[];
-  messages: AgentTeamMessageRecord[];
-  checkpointRound?: number;
-}
-
-export type AgentTeamPhase = 'briefing' | 'plan_review' | 'starting' | 'executing';
-
-export interface AgentTeamPlanTeammate {
-  name: string;
-  agentType?: string;
-}
-
-export interface AgentTeamPlanTask {
-  title: string;
-  description: string;
-  ownerName?: string;
-  deps: string[];
-  /** File or directory paths this task may create or modify. */
-  scope?: string[];
-  /** Mechanical verification command, or 'manual'. */
-  verify?: string;
-}
-
-export interface AgentTeamPlanDraft {
-  summary: string;
-  teammates: AgentTeamPlanTeammate[];
-  tasks: AgentTeamPlanTask[];
-  /** Team-level command proving the whole deliverable works together. */
-  integrationVerify?: string;
-  sourceAgentId?: string;
-  createdAt: number;
-  updatedAt: number;
-}
-
-/** Liveness of an agent's PTY session as reported by the main process. */
-export type AgentTeamSessionHealth = 'live' | 'queued' | 'dead' | 'missing';
-
-export interface AgentTeamSnapshot {
-  workspaceId: string;
-  frameNodeId?: string;
-  phase: AgentTeamPhase;
-  pendingPlan?: AgentTeamPlanDraft;
-  approvedPlan?: AgentTeamPlanDraft;
-  runtime: AgentTeamRuntimeSnapshot;
-  /** Per-agent session liveness, keyed by agent id. */
-  sessions?: Record<string, AgentTeamSessionHealth>;
-}
-
-export interface AgentTeamsApi {
-  create: (input: {
-    workspaceId: string;
-    name: string;
-    goal: string;
-    cwd?: string;
-    leadName?: string;
-    leadAgentType?: string;
-    teammateNames?: string[];
-    teammateAgentType?: string;
-    x?: number;
-    y?: number;
-  }) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  list: (workspaceId: string) => Promise<{ ok: boolean; teams?: AgentTeamSnapshot[]; error?: string }>;
-  snapshot: (workspaceId: string, teamId: string) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  addAgent: (input: {
-    workspaceId: string;
-    teamId: string;
-    name: string;
-    role: 'lead' | 'teammate';
-    agentType?: string;
-    cwd?: string;
-  }) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  briefLead: (
-    workspaceId: string,
-    teamId: string,
-    content: string,
-  ) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  confirmPlan: (
-    workspaceId: string,
-    teamId: string,
-  ) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  updatePlanTeammate: (
-    workspaceId: string,
-    teamId: string,
-    teammateName: string,
-    agentType: string,
-  ) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  createTask: (input: {
-    workspaceId: string;
-    teamId: string;
-    title: string;
-    description: string;
-    ownerAgentId?: string;
-    ownerName?: string;
-    deps?: string[];
-    depRefs?: string[];
-    scope?: string[];
-    verify?: string;
-    dispatch?: boolean;
-  }) => Promise<{ ok: boolean; runtime?: AgentTeamRuntimeSnapshot; error?: string }>;
-  advanceRound: (
-    workspaceId: string,
-    teamId: string,
-  ) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  finalizeFromCheckpoint: (
-    workspaceId: string,
-    teamId: string,
-  ) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  updateTask: (
-    workspaceId: string,
-    teamId: string,
-    taskId: string,
-    title?: string,
-    description?: string,
-  ) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  dispatch: (workspaceId: string, teamId: string) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  pause: (workspaceId: string, teamId: string) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  resume: (workspaceId: string, teamId: string) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  prepareAgentAutoResume: (
-    workspaceId: string,
-    teamId: string,
-    agentId: string,
-  ) => Promise<{ ok: boolean; canResume?: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  delete: (
-    workspaceId: string,
-    teamId: string,
-  ) => Promise<{ ok: boolean; deletedNodeIds?: string[]; error?: string }>;
-  completeTask: (
-    workspaceId: string,
-    teamId: string,
-    taskId: string,
-    result?: string,
-  ) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  openGate: (input: {
-    workspaceId: string;
-    teamId: string;
-    agentId?: string;
-    taskId?: string;
-    reason: string;
-    prompt: string;
-  }) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  answerGate: (workspaceId: string, gateId: string, answer: string) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  interruptAgent: (input: {
-    workspaceId: string;
-    teamId: string;
-    agentId: string;
-    mode: 'soft' | 'ctrl-c' | 'abort';
-    reason?: string;
-  }) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-  sendInput: (
-    workspaceId: string,
-    teamId: string,
-    agentId: string,
-    content: string,
-  ) => Promise<{ ok: boolean; snapshot?: AgentTeamSnapshot; error?: string }>;
-}
-
-export interface CodexSessionIndexEntry {
-  id: string;
-  threadName?: string;
-  updatedAt: string;
-}
-
-export interface CodexThreadMatch {
-  id: string;
-  cwd?: string;
-  title?: string;
-  updatedAtMs?: number;
-}
-
-export interface CodexSessionsApi {
-  list: (
-    payload?: { updatedAfter?: string },
-  ) => Promise<{ ok: boolean; sessions?: CodexSessionIndexEntry[]; error?: string }>;
-  findByMarker: (
-    payload: { marker: string; updatedAfterMs?: number; cwd?: string },
-  ) => Promise<{
-    ok: boolean;
-    session?: CodexThreadMatch;
-    ambiguous?: boolean;
-    error?: string;
-  }>;
-}
-
-export interface CanvasWorkspaceApi {
-  version: string;
-  appInfo: AppInfoApi;
-  pluginFlags: Record<string, boolean>;
-  pty: {
-    spawn: (
-      id: string,
-      cols?: number,
-      rows?: number,
-      cwd?: string,
-      workspaceId?: string,
-      env?: Record<string, string | undefined>
-    ) => Promise<{ ok: boolean; pid?: number; error?: string; reused?: boolean }>;
-    write: (id: string, data: string) => void;
-    resize: (id: string, cols: number, rows: number) => void;
-    kill: (id: string) => void;
-    getCwd: (id: string) => Promise<{ ok: boolean; cwd?: string | null }>;
-    checkCommand: (command: string) => Promise<{ ok: boolean; available: boolean; path?: string; error?: string }>;
-    onData: (id: string, callback: (data: string) => void) => () => void;
-    onExit: (id: string, callback: (exitCode: number) => void) => () => void;
-  };
-  store: {
-    save: (
-      id: string,
-      data: unknown
-    ) => Promise<{ ok: boolean; error?: string }>;
-    load: (
-      id: string
-    ) => Promise<{ ok: boolean; data?: CanvasSaveData | null; error?: string }>;
-    list: () => Promise<{ ok: boolean; ids?: string[]; error?: string }>;
-    delete: (id: string) => Promise<{ ok: boolean; error?: string }>;
-    getDir: (id: string) => Promise<{ ok: boolean; dir?: string; error?: string }>;
-    exportWorkspace: (
-      id: string,
-      name: string
-    ) => Promise<{ ok: boolean; canceled?: boolean; filePath?: string; fileCount?: number; error?: string }>;
-    importWorkspace: () => Promise<{
-      ok: boolean;
-      canceled?: boolean;
-      workspaceId?: string;
-      workspaceName?: string;
-      fileCount?: number;
-      error?: string;
-    }>;
-    /**
-     * Returns workspaces whose canvas.json was clobbered by a v1-unaware
-     * writer (signature: v1-shape canvas.json with node ids that overlap
-     * existing nodes/<id>.json files). Renderer surfaces sticky alerts
-     * for each; recovery is via `canvas-cli restore`.
-     */
-    listPollutedWorkspaces: () => Promise<{
-      ok: boolean;
-      polluted?: Array<{ workspaceId: string; conflictingNodeIds: string[] }>;
-      error?: string;
-    }>;
-    watchWorkspace: (workspaceId: string) => Promise<{ ok: boolean }>;
-    onExternalUpdate: (
-      callback: (event: {
-        workspaceId: string;
-        nodeIds: string[];
-        kind?: "create" | "update" | "delete";
-        source: string;
-      }) => void
-    ) => () => void;
-    /**
-     * Subscribe to canvas storage migration progress events.
-     *
-     * `errorKind` distinguishes a critical data-integrity event
-     * (`'pollution'` — a v1-unaware writer clobbered canvas.json over a
-     * v2 workspace) from a generic migration hiccup; the renderer should
-     * surface the former as a sticky alert and the latter as a short
-     * toast.
-     */
-    onMigrationProgress: (
-      callback: (event: {
-        workspaceId: string;
-        phase:
-        | "starting"
-        | "backup"
-        | "split-nodes"
-        | "commit"
-        | "done"
-        | "error";
-        current?: number;
-        total?: number;
-        message?: string;
-        errorKind?: "pollution" | "other";
-        conflictingNodeIds?: string[];
-      }) => void
-    ) => () => void;
-  };
-  workspaceNodes: {
-    list: (workspaceId: string) => Promise<{
-      ok: boolean;
-      nodes?: WorkspaceNodeListItem[];
-      tags?: KnowledgeTagDefinition[];
-      error?: string;
-    }>;
-    read: (workspaceId: string, nodeId: string) => Promise<{
-      ok: boolean;
-      node?: WorkspaceNodeRecord | null;
-      error?: string;
-    }>;
-    tags: () => Promise<{
-      ok: boolean;
-      tags?: KnowledgeTagDefinition[];
-      error?: string;
-    }>;
-    upsertTag: (tag: { id?: string; name: string; description?: string }) => Promise<{
-      ok: boolean;
-      tag?: KnowledgeTagDefinition;
-      error?: string;
-    }>;
-    updateTags: (workspaceId: string, nodeId: string, tags: string[]) => Promise<{
-      ok: boolean;
-      node?: WorkspaceNodeRecord | null;
-      error?: string;
-    }>;
-    update: (workspaceId: string, nodeId: string, patch: Partial<WorkspaceNodeRecord>) => Promise<{
-      ok: boolean;
-      node?: WorkspaceNodeRecord | null;
-      error?: string;
-    }>;
-    /** Fires when workspace-node metadata (tags / properties) changes in the main process. */
-    onChange: (
-      callback: (event: { workspaceIds: string[]; source: 'canvas-agent' | 'renderer' }) => void,
-    ) => () => void;
-  };
-  file: FileApi;
-  dialog: DialogApi;
-  skills: SkillsApi;
-  canvasSkills: CanvasSkillsApi;
-  canvasMcp: CanvasMcpApi;
-  experimental: ExperimentalApi;
-  channelConfig: ChannelConfigApi;
-  builtInTools: BuiltInToolsConfigApi;
-  model: CanvasModelApi;
-  promptProfile: PromptProfileApi;
-  agent: AgentApi;
-  codexSessions: CodexSessionsApi;
-  agentTeams: AgentTeamsApi;
-  iframe: IframeApi;
-  llm: LlmApi;
-  artifacts: ArtifactsApi;
-  shell: ShellApi;
-  link: LinkApi;
-  web: WebApi;
-  plugin: PluginBridge;
-}
-
-export interface ShellApi {
-  openExternal: (url: string) => Promise<{ ok: boolean; error?: string }>;
-}
-
-export interface AppInfoApi {
-  getInfo: () => Promise<AppInfoResult>;
-  checkForUpdates: () => Promise<UpdateCheckResult>;
-}
-
-export interface AppInfoResult {
-  ok: boolean;
-  version?: string;
-  updateManifestUrl?: string;
-  downloadUrl?: string;
-  error?: string;
-}
-
-export interface UpdateManifestSummary {
-  version: string;
-  releasedAt?: string;
-  downloadUrl: string;
-  notes?: string | {
-    zh?: string;
-    en?: string;
-  };
-}
-
-export type UpdateCheckResult =
-  | {
-      ok: true;
-      currentVersion: string;
-      updateAvailable: boolean;
-      latest: UpdateManifestSummary;
-    }
-  | {
-      ok: false;
-      currentVersion?: string;
-      error?: string;
-    };
-
-export interface ExperimentalFeatureDef {
-  id: string;
-  label: string;
-  description: string;
-  defaultEnabled: boolean;
-}
-
-/**
- * Pushed from main when a flag toggle kicks off a background tooling install
- * (e.g. the Agent Teams skill + CLI). Delivered over the
- * `experimental:tooling-status` channel once the install settles.
- */
-export interface ToolingInstallStatus {
-  /** The experimental flag id that triggered the install. */
-  feature: string;
-  ok: boolean;
-  skillsInstalled: boolean;
-  cliInstalled: boolean;
-  cliError?: string | null;
-  manualCommand?: string | null;
-}
-
-export interface ExperimentalApi {
-  list: () => Promise<{
-    ok: boolean;
-    features?: ExperimentalFeatureDef[];
-    values?: Record<string, boolean>;
-    path?: string;
-    error?: string;
-  }>;
-  set: (id: string, enabled: boolean) =>
-    Promise<{ ok: boolean; values?: Record<string, boolean>; error?: string }>;
-  reset: () => Promise<{ ok: boolean; values?: Record<string, boolean>; error?: string }>;
-  reloadWindow: () => Promise<{ ok: boolean; error?: string }>;
-  /** Subscribe to background tooling-install results. Returns an unsubscribe fn. */
-  onToolingStatus: (cb: (status: ToolingInstallStatus) => void) => () => void;
-}
-
-export interface ChannelConfigStatus {
-  path: string;
-  feishu: {
-    appId?: string;
-    secretPresent: boolean;
-    defaultWorkspaceId?: string;
-    appIdFromEnv: boolean;
-    secretFromEnv: boolean;
-    defaultWorkspaceFromEnv: boolean;
-  };
-}
-
-export interface SetFeishuConfigInput {
-  appId?: string;
-  /** New secret to store. Empty/omitted leaves the existing secret untouched. */
-  appSecret?: string;
-  defaultWorkspaceId?: string;
-  /** When true, remove the stored secret. */
-  clearSecret?: boolean;
-}
-
-export interface ChannelConfigApi {
-  status: () => Promise<{ ok: boolean; status?: ChannelConfigStatus; error?: string }>;
-  setFeishu: (
-    input: SetFeishuConfigInput,
-  ) => Promise<{ ok: boolean; status?: ChannelConfigStatus; error?: string }>;
-  clearFeishu: () => Promise<{ ok: boolean; status?: ChannelConfigStatus; error?: string }>;
-  /** Relaunch the app so credential / flag changes take effect. */
-  relaunch: () => Promise<{ ok: boolean; error?: string }>;
-}
-
-export interface LinkApi {
-  /** Subscribe to URLs intercepted from embedded webviews / iframes. Returns unsubscribe fn. */
-  onOpen: (callback: (data: { url: string }) => void) => () => void;
-}
-
-export interface LlmApi {
-  /** One-shot generation — returns full HTML when complete. */
-  generateHTML: (prompt: string) => Promise<{ ok: boolean; html?: string; error?: string }>;
-  /** Start a streaming generation — returns a requestId to subscribe to deltas. */
-  streamHTML: (prompt: string) => Promise<{ ok: boolean; requestId?: string; error?: string }>;
-  /** Subscribe to incremental text chunks during streaming generation. Returns unsubscribe fn. */
-  onHTMLDelta: (requestId: string, callback: (delta: string) => void) => () => void;
-  /** Subscribe to generation completion. Returns unsubscribe fn. */
-  onHTMLComplete: (requestId: string, callback: (result: { ok: boolean; html?: string; error?: string }) => void) => () => void;
-}
-
-export type WebReadStrategy = 'auto' | 'dom' | 'a11y' | 'screenshot';
-
-export interface WebReadInput {
-  workspaceId: string;
-  nodeId: string;
-  strategy?: WebReadStrategy;
-  maxChars?: number;
-  sparseThreshold?: number;
-}
-
-export type WebReadResult =
-  | { ok: true; nodeId: string; strategy: 'dom'; text: string; title: string; url: string }
-  | { ok: true; nodeId: string; strategy: 'a11y'; text: string }
-  | { ok: true; nodeId: string; strategy: 'screenshot'; imagePath: string }
-  | { ok: false; nodeId: string; strategy: WebReadStrategy; error: string };
-
-export interface WebApi {
-  read: (payload: WebReadInput) => Promise<WebReadResult>;
-}
-
-/**
- * An LLM-generated visual product owned by a workspace.
- *
- * Three formation paths share the same data shape:
- *  1. `visual_render` tool → inline-only, NOT stored here (lives in chat).
- *  2. `artifact_create` tool → stored, versioned, surfaced via chat card + drawer.
- *  3. User clicks "Save as artifact" on an inline visual → promoted into the store.
- *
- * Type values are open to extension; v1 covers `html` only.
- */
-export type ArtifactType = 'html' | 'svg' | 'mermaid';
-
-export interface ArtifactVersion {
-  id: string;
-  content: string;
-  /** Optional prompt that produced this version — useful as a "diff" hint. */
-  prompt?: string;
-  createdAt: number;
-}
-
-export interface Artifact {
-  id: string;
-  workspaceId: string;
-  type: ArtifactType;
-  title: string;
-  versions: ArtifactVersion[];
-  /** Index into `versions`. Always `versions.length - 1` after a create/update,
-   * but a separate field so the renderer can switch views without mutating data. */
-  currentVersionId: string;
-  /** When set, the artifact is currently mirrored as this canvas node. */
-  pinnedNodeId?: string;
-  /** Origin trace — where the artifact came from. */
-  source?: {
-    sessionId?: string;
-    /** Index of the assistant message in that session that produced it. */
-    messageIndex?: number;
-    origin?: 'agent_tool' | 'inline_promotion' | 'iframe_ai_tab';
-  };
-  createdAt: number;
-  updatedAt: number;
-}
-
-export interface ArtifactsApi {
-  list: (workspaceId: string) => Promise<{ ok: boolean; artifacts?: Artifact[]; error?: string }>;
-  get: (
-    workspaceId: string,
-    artifactId: string,
-  ) => Promise<{ ok: boolean; artifact?: Artifact; error?: string }>;
-  create: (
-    workspaceId: string,
-    input: {
-      type: ArtifactType;
-      title: string;
-      content: string;
-      prompt?: string;
-      source?: Artifact['source'];
-    },
-  ) => Promise<{ ok: boolean; artifact?: Artifact; error?: string }>;
-  /** Append a new version to an existing artifact, becoming the current one. */
-  addVersion: (
-    workspaceId: string,
-    artifactId: string,
-    input: { content: string; prompt?: string },
-  ) => Promise<{ ok: boolean; artifact?: Artifact; error?: string }>;
-  /** Mutate metadata only (title, currentVersionId, pinnedNodeId). */
-  update: (
-    workspaceId: string,
-    artifactId: string,
-    patch: Partial<Pick<Artifact, 'title' | 'currentVersionId' | 'pinnedNodeId'>>,
-  ) => Promise<{ ok: boolean; artifact?: Artifact; error?: string }>;
-  delete: (
-    workspaceId: string,
-    artifactId: string,
-  ) => Promise<{ ok: boolean; error?: string }>;
-  /** Create an iframe canvas node bound to this artifact (mode='artifact'). */
-  pinToCanvas: (
-    workspaceId: string,
-    artifactId: string,
-    placement?: { x?: number; y?: number; width?: number; height?: number; title?: string },
-  ) => Promise<{ ok: boolean; nodeId?: string; artifact?: Artifact; error?: string }>;
-  /** Fires when an artifact is created, updated, or deleted in the main process. */
-  onChange: (
-    callback: (event: {
-      workspaceId: string;
-      artifactId: string;
-      kind: 'create' | 'update' | 'delete';
-    }) => void,
-  ) => () => void;
-}
-
-export interface IframeApi {
-  registerWebview: (
-    workspaceId: string,
-    nodeId: string,
-    webContentsId: number,
-  ) => Promise<{ ok: boolean }>;
-  unregisterWebview: (
-    workspaceId: string,
-    nodeId: string,
-  ) => Promise<{ ok: boolean }>;
-}
+import type { CanvasWorkspaceApi } from './types/workspace-api';
+
+export type * from './types/agent-chat';
+export type * from './types/agent-teams';
+export type * from './types/app-info';
+export type * from './types/artifacts';
+export type * from './types/canvas';
+export type * from './types/channel-config';
+export type * from './types/codex-sessions';
+export type * from './types/experimental';
+export type * from './types/files';
+export type * from './types/iframe';
+export type * from './types/link';
+export type * from './types/llm';
+export type * from './types/models';
+export type * from './types/settings-config';
+export type * from './types/shell';
+export type * from './types/web';
+export type * from './types/workspace-api';
 
 declare global {
   interface Window {
