+++ apps/canvas-workspace/src/renderer/src/types/app-info.ts
@@ -1,0 +1,11 @@
+import type {
+  AppInfoResult,
+  UpdateCheckResult,
+} from '../../../shared/app-info';
+
+export type * from '../../../shared/app-info';
+
+export interface AppInfoApi {
+  getInfo: () => Promise<AppInfoResult>;
+  checkForUpdates: () => Promise<UpdateCheckResult>;
+}
