+++ apps/canvas-workspace/src/renderer/src/types/canvas.ts
@@ -1,0 +1,1 @@
+export type * from '../../../shared/canvas';
