+++ apps/canvas-workspace/src/renderer/src/types/iframe.ts
@@ -1,0 +1,11 @@
+export interface IframeApi {
+  registerWebview: (
+    workspaceId: string,
+    nodeId: string,
+    webContentsId: number,
+  ) => Promise<{ ok: boolean }>;
+  unregisterWebview: (
+    workspaceId: string,
+    nodeId: string,
+  ) => Promise<{ ok: boolean }>;
+}
