+++ apps/canvas-workspace/src/renderer/src/types/link.ts
@@ -1,0 +1,4 @@
+export interface LinkApi {
+  /** Subscribe to URLs intercepted from embedded webviews / iframes. Returns unsubscribe fn. */
+  onOpen: (callback: (data: { url: string }) => void) => () => void;
+}
