+++ apps/canvas-workspace/src/renderer/src/types/shell.ts
@@ -1,0 +1,3 @@
+export interface ShellApi {
+  openExternal: (url: string) => Promise<{ ok: boolean; error?: string }>;
+}
