+++ apps/canvas-workspace/src/renderer/src/types/web.ts
@@ -1,0 +1,10 @@
+import type {
+  WebReadInput,
+  WebReadResult,
+} from '../../../shared/web';
+
+export type * from '../../../shared/web';
+
+export interface WebApi {
+  read: (payload: WebReadInput) => Promise<WebReadResult>;
+}
