+++ apps/canvas-workspace/src/shared/codex-sessions.ts
@@ -1,0 +1,12 @@
+export interface CodexSessionIndexEntry {
+  id: string;
+  threadName?: string;
+  updatedAt: string;
+}
+
+export interface CodexThreadMatch {
+  id: string;
+  cwd?: string;
+  title?: string;
+  updatedAtMs?: number;
+}
