+++ apps/canvas-workspace/src/shared/files.ts
@@ -1,0 +1,5 @@
+export interface DirEntry {
+  name: string;
+  type: 'file' | 'dir';
+  children?: DirEntry[];
+}
